requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncmy==0.2.8
asyncpg==0.28.0
attrs==23.2.0
bcrypt==4.0.1
black==23.10.1
blinker==1.7.0
Brotli==1.1.0
brotli-asgi==1.4.0
bunnycdnpython==0.0.8
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
cryptography==3.4.8
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
fastapi-asyncapi==0.1.0
fastapi-cache2==0.2.1
fastapi-jwt-auth==0.5.0
Flask==3.0.0
Flask-WTF==1.2.1
flupy==1.2.0
frozenlist==1.4.1
greenlet==3.0.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
mypy==1.6.1
mypy-extensions==1.0.0
nanoid==2.0.0
oauthlib==3.2.2
orjson==3.9.9
packaging==23.2
parse==1.19.1
passlib==1.7.4
pathspec==0.11.2
pendulum==3.0.0
platformdirs==3.11.0
pluggy==1.4.0
pyaes==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
Pygments==2.17.2
pyinstrument==4.6.2
PyJWT==1.7.1
PySocks==1.7.1
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-dotenv==0.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
sentry-sdk==1.32.0
shellingham==1.4.0
six==1.16.0
sniffio==1.3.0
sqladmin==0.16.1
SQLAlchemy==2.0.29
starlette==0.27.0
telegram==0.0.1
TgCrypto==1.2.5
time-machine==2.13.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2024.1
ujson==5.8.0
ulid-py==1.1.0
urllib3==2.0.7
uvicorn==0.23.2
watchfiles==0.21.0
watchgod==0.8.2
websockets==12.0
Werkzeug==3.0.1
WTForms==3.1.1
yadisk==2.1.0
yarl==1.9.4
